+++ CI/utility_functions.bash
@@ -196,12 +196,15 @@ run_integration_tests() {
     sudo -u docker --login bash -c "
         PATH=/opt/sarus/default/bin:\$PATH PYTHONPATH=/sarus-source/CI/src:\$PYTHONPATH \
         CMAKE_INSTALL_PREFIX=/opt/sarus/default HOME=/home/docker \
-        pytest -v -m 'not asroot' /sarus-source/CI/src/integration_tests/"
+        pytest -v -m 'not asroot' /sarus-source/CI/src/integration_tests/" # TIP: Add -s --last-failed to pytest to get more output from failed tests.
     fail_on_error "Python integration tests failed"
     echo "Successfully run integration tests with user=docker"
 
     echo "Running integration tests with user=root"
-    sudo --login bash -c "PATH=/opt/sarus/default/bin:\$PATH PYTHONPATH=/sarus-source/CI/src:$PYTHONPATH CMAKE_INSTALL_PREFIX=/opt/sarus/default HOME=/home/docker pytest -v -m asroot /sarus-source/CI/src/integration_tests/"
+    sudo --login bash -c "
+        PATH=/opt/sarus/default/bin:\$PATH PYTHONPATH=/sarus-source/CI/src:$PYTHONPATH \
+        CMAKE_INSTALL_PREFIX=/opt/sarus/default HOME=/home/docker \
+        pytest -v -s -m asroot /sarus-source/CI/src/integration_tests/"
     fail_on_error "Python integration tests as root failed"
     echo "Successfully run integration tests with user=root"
 }