+++ src/hooks/slurm_global_sync/main.cpp
@@ -17,6 +17,7 @@
 int main(int argc, char* argv[]) {
     try {
         auto hook = sarus::hooks::slurm_global_sync::Hook{};
+        hook.dropPrivileges();
         hook.loadConfigs();
         hook.performSynchronization();
     } catch(const sarus::common::Error& e) {