+++ src/runtime/mount_utilities.hpp
@@ -28,7 +28,8 @@ namespace runtime {
 
 void validateMountSource(const boost::filesystem::path&);
 void validateMountDestination(const boost::filesystem::path& destination, const common::Config& config);
-bool isPathOnAllowedDevice(const boost::filesystem::path& path, const common::Config& config);
+void validateMountDestination(const boost::filesystem::path& destination, const boost::filesystem::path& bundleDir, const boost::filesystem::path& rootfsDir);
+bool isPathOnAllowedDevice(const boost::filesystem::path& path, const boost::filesystem::path& bundleDir, const boost::filesystem::path& rootfsDir);
 dev_t getDevice(const boost::filesystem::path& path);
 void bindMount(const boost::filesystem::path& from, const boost::filesystem::path& to, unsigned long flags=0);
 void loopMountSquashfs(const boost::filesystem::path& image, const boost::filesystem::path& mountPoint);